refactor(debug-info): move DebugInfo into separate component
