+++ source/components/organisms/DebugInfo/DebugInfo.types.ts
@@ -1,0 +1,5 @@
+import type { DebugInfoCategory } from "../../../helpers/debugInfo/debugInfo.types";
+
+export interface Props {
+  debugInfo?: DebugInfoCategory[];
+}
