+++ source/screens/debugInfoScreen/DebugInfoScreen.types.ts
@@ -1,0 +1,5 @@
+export interface Props {
+  navigation: {
+    navigate: (to: string) => void;
+  };
+}
